ORCA-142: Reimplement coveralls support from PR#51.

+++ bin/travis/before_install.sh
@@ -23,8 +23,8 @@ echo "$TRAVIS_NODE_VERSION"
 # Display the Yarn version.
 yarn --version
 
-# Disable Xdebug.
-phpenv config-rm xdebug.ini
+# Disable Xdebug except on code coverage jobs.
+[[ "$ORCA_JOB" = "ISOLATED_RECOMMENDED_COVERAGE" ]] || phpenv config-rm xdebug.ini
 
 {
   # Remove PHP memory limit.
@@ -36,6 +36,7 @@ phpenv config-rm xdebug.ini
 } >> "$HOME/.phpenv/versions/$(phpenv version-name)/etc/conf.d/travis.ini"
 
 # Install the PECL YAML parser for strict YAML parsing.
+pecl channel-update pecl.php.net
 yes | pecl install yaml
 
 # Display PHP information.
